requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
httpx>=0.28,<1.0
numpy>=1.26,<3.0
pydantic>=2.7,<3.0
python-dotenv>=1.0,<2.0
# Warp
shapely>=2.0,<3.0
mapbox-vector-tile>=2.0,<3.0
# Meetup
h3>=4.1,<5.0